refactor: [GNS-354] Extract StoragePrice as common
