+++ src/common/react-query/statistics/use-get-storage-price.ts
@@ -3,16 +3,16 @@ import { UseQueryOptions } from "react-query";
 import { QUERY_KEY } from "@/common/react-query/query-keys";
 import { useServiceProvider } from "@/common/hooks/provider/use-service-provider";
 import { useApiRepositoryQuery } from "@/common/react-query/hoc/api";
-import { API_REPOSITORY_KEY } from "@/common/values/query.constant";
+import { RPC_REPOSITORY_KEY } from "@/common/values/query.constant";
 import { Amount } from "@/types";
 
 export const useGetStoragePrice = (options?: UseQueryOptions<Amount | null, Error, Amount | null>) => {
-  const { apiStatisticsRepository } = useServiceProvider();
+  const { chainRepository } = useServiceProvider();
 
   return useApiRepositoryQuery(
     [QUERY_KEY.getStoragePrice],
-    apiStatisticsRepository,
-    API_REPOSITORY_KEY.STATISTICS_REPOSITORY,
+    chainRepository,
+    RPC_REPOSITORY_KEY.CHAIN_REPOSITORY,
     repository => repository.getStoragePrice(),
     options,
   );
